+++ instrumentation/opentelemetry-instrumentation-django/tests/test_middleware_asgi.py
@@ -20,7 +20,6 @@ from django import VERSION, conf
 from django.http import HttpRequest, HttpResponse
 from django.test import SimpleTestCase
 from django.test.utils import setup_test_environment, teardown_test_environment
-from django.urls import re_path
 
 from opentelemetry.instrumentation.django import (
     DjangoInstrumentor,
@@ -54,7 +53,13 @@ from .views import (
     async_traced_template,
 )
 
+DJANGO_2_0 = VERSION >= (2, 0)
 DJANGO_3_1 = VERSION >= (3, 1)
+
+if DJANGO_2_0:
+    from django.urls import re_path
+else:
+    from django.conf.urls import url as re_path
 
 urlpatterns = [
     re_path(r"^traced/", async_traced),
@@ -344,7 +349,7 @@ class TestMiddlewareAsgi(SimpleTestCase, TestBase):
     async def test_trace_response_headers(self):
         response = await self.async_client.get("/span_name/1234/")
 
-        self.assertNotIn("Server-Timing", response.headers)
+        self.assertFalse(response.has_header("Server-Timing"))
         self.memory_exporter.clear()
 
         set_global_response_propagator(TraceResponsePropagator())
@@ -352,12 +357,12 @@ class TestMiddlewareAsgi(SimpleTestCase, TestBase):
         response = await self.async_client.get("/span_name/1234/")
         span = self.memory_exporter.get_finished_spans()[0]
 
-        self.assertIn("traceresponse", response.headers)
+        self.assertTrue(response.has_header("traceresponse"))
         self.assertEqual(
-            response.headers["Access-Control-Expose-Headers"], "traceresponse",
+            response["Access-Control-Expose-Headers"], "traceresponse",
         )
         self.assertEqual(
-            response.headers["traceresponse"],
+            response["traceresponse"],
             "00-{}-{}-01".format(
                 format_trace_id(span.get_span_context().trace_id),
                 format_span_id(span.get_span_context().span_id),